Modify mounting paths
When we create mounting paths for devices we have same paths if we create
them on 2 servers. For example for 2 servers and 2 volumes paths look like
this:
- /mnt/vdb
- /mnt/vdc
- /mnt/vdb
- /mnt/vdc
In this patch we add one more directory - id of server to this paths,
that when we have unique patch for each. After this patch paths will look
like this:
- /mnt/226575324/vdb
- /mnt/226575324/vdc
- /mnt/356104023/vdb
- /mnt/356104023/vdc

Change-Id: I7e4b2f68d987cd4d303915062984ce2125667db6