Allow visibility of external subnet as shared ones
The Neutron network can have the attribute "external". If a
network is external, can be seen by other projects, same as
when it is shared. However the subnets cannot be seen now.
This issue is being solved in LP#2051831, in [1].

This patch changes the logic implemented in [2], where the
subnets belonging to an external network where hidden to
a non-admin user of other project.

[1]https://review.opendev.org/c/openstack/neutron/+/907313
[2]https://review.opendev.org/c/openstack/tempest/+/484960

Related-Bug: #2051831
Change-Id: I58b2a1ddf9b0fe85d96617863d121aebb7d81e58