Switch the default disk and container formats to bare/qcow2
The first item of image.disk_formats and image.container_formats
tempest options is used by default if the options are not
specified by the person runnig tempest.
The most likely format to be used and supported are bare and qcow2,
so let's switch to those as default options.

Also, move raw to the second place of disk_formats because
a specific test (test_list_images_param_disk_format from the
tempest.api.image.v2.test_images.ListUserImagesTest class)
requires an image in raw format, but its initializer only
creates a subset of the images (only the first 6 (!) items
of all combinations of container_formats and disk_formats)
which means that raw should be in the first 6 combinations.

Change-Id: Iab3e5205e2ec6cd5a253715ff87853b2805e4819