Add tests for creating servers with (anti-)affinity
Currently there are only two tests for server affinity and
anti-affinity and both of them use the /servers multi-create API and so
are not covering cases where server(s) in a group are already present
on hosts when a new group member is being scheduled.

This adds some additional tests for cases of already present server
group members and for soft (anti-)affinity.

Change-Id: I84a9f9da666e51fad8b5bf4e214e1027b2f56766