Add configurable hostname pattern to filter hosts
Currently when aggregating potential target hosts via
get_host_other_than, it's hostname filtering only addresses the suffix
'-ironic'. In the event that pattern is located within the hostname e.g.
nova-cell1-compute-ironic-compute-0, then it will still be used as a
target candidate. Update tempest to allow for a configurable hostname
pattern to filter on when searching for candidates.

Also updated additional testcases that are still using the suffix only
check when removing guests with ironic in the name.

Change-Id: If5bab817e04412512186be2195cd2437bd310980