Test image formats
This implements testing of image formats based on the sample images
in os-test-images. We should be able to assert that formats we accept
are allowed by glance and formats that we do not are rejected.

Note that glance currently does not do enough of this validation, so
not nearly enough of the unusable samples are currently rejected,
but this will serve as a base from which to start implementing and
testing that in glance.

This adds testscenarios as a dependency and uses that utility along
with the load_tests() protocol to generate test scenarios from the
manifest file. This results in separate tests for each image format,
without us needing to manually add those cases (and without the risk
of missing some because we don't).

Depends-On: https://review.opendev.org/c/openstack/devstack/+/925425
Change-Id: I4536b6b36b23071447ea8efbfcd2b3a313414034