Allow nova vmdk-footer test to pass
Nova originally did not support vmdk-with-footer and the test asserted
that fact. Now we need to fix nova and have this test work with both
fixed and non-fixed code. Since the tests are not split by image
format, we can easily skip in the case where an image is marked as
usable, but fails in a known way, which will allow us to pass the
job with and without a fixed nova.

Change-Id: I71268fa915055300e1cb42e592bc62a1f276d733