Wait for all instance ports to become ACTIVE
get_server_port_id_and_ip4() gets a list of neutron ports
for an instance, but it could be one or more of those have
not completed provisioning at the time of the call, so are
still marked DOWN.

Wait for all ports to become active since it could just be
neutron has not completed its work yet.

Added new waiter function and tests to verify it worked.

Closes-bug: #2079831
Change-Id: I758e5eeb8ab05e79d6bdb2b560aa0f9f38c5992c