Use stable constraint in tox to release new tag for 2024.2
We are going to release Tempest new tag to declare the start of
support for stable/2024.2. So that new tag use stable/2024.2
constraint in the tox env.

For example: if anyone use Tempest new tag in future say
1 year later then tox env also should use the 2024.2 contstraint for
compatibility instead master.

Also, making ironic-tempest-bios-ipmi-direct-tinyipa as
non voting to merge his change.

These need to be reverted once release is done.

Change-Id: Iac4607281c40e7583e181ab1bf04331a68d19464