Stop violating image disk_format rules
Glance is going to start enforcing that if you upload something and
claim it is a different format than the content, the upload will fail.
Tempest currently just almost universally uses disk_formats[0],
despite not actually uploading content that matches that format. In
most cases it is using random data, which means the format it uses
should be "raw". This converts those cases to use raw, and does so
assuming raw is always available (as many other tests already do).

Change-Id: I06d66b969306b53cd177c2d95c23762db3bc72ea
Needed-By: https://review.opendev.org/c/openstack/glance/+/930492