Avoid image format tests if glance is enforcing
If glance refuses to allow bad images to be uploaded, then we cannot
even register them to then test against nova or other services. Gate
these tests on a knob indicating whether or not glance will allow
them to be uploaded. The ListUserImagesTest class depends entirely on
being able to upload random data as every format/container, which is
fundamentally not something we can do with glance enforcing content
rules.

This makes the tempest-slow-py3 job disable glance image format
inspection so that we can get the compute tests to run.

Depends-On: https://review.opendev.org/c/openstack/devstack/+/931026
Change-Id: I05bb233cf1482895b96396852a15532295d3829b