workaround: Wait for instance.host=None when shelving
There is a race in Nova where an instance being shelved is set to the
SHELVED_OFFLOADED vm_state before the instance.host is cleared. It
appears to be non-trivial to fix given the comments in that code
explictly say the updates have the be in that order.

This works around the issue by waiting until the instance.host is None
before considering the shelve offload to be completed.

Related-Bug: #2045785

Change-Id: Ia45247a7aa14eb5f0038d4512a0b4ebe6af5a573