Remove python 3.8 support
In this cycle, oslo library are dropping the
python3.8 support. Without oslo libs, Tempest
cannot support the python 3.8. Tempest can be run
in virtual env and upstream CI also test the stable
supported branches in virtual env. If anyone want to
test the cloud on python 3.8, they can still run the
Tempest in virtual env python 3.9.

This patch removes the support of python 3.8 and its testing.

Change-Id: I3f540da045c5e89a3f09ce2cf94877c1560f3436