Check only ports owned by the dhcp service in the dhcp scheduler tests
In the DHCPAgentSchedulersTestJSON.test_dhcp_port_status_active test
it was checked if all ports from the network are ACTIVE. This is
typically correct but in case of the e.g. ML2/OVN backend with deployed
Neutron DHCP agent (which is rare but still possible, supported and
required in some corner use cases) this test was failing as in addition
to the port with device owner ``network:dhcp`` there was also
``network:distributed`` port created by the ovn driver and this port is
DOWN which is normal.

To avoid such false failure of this test this patch adds filtering only
for the ports with the owner like: ``network:dhcp``.

Change-Id: Id53a4b4dffe5ec18bbd9eaf2a9e74175b4dd87bb