Allow glance support for GPT images
This tweaks the image format tests to allow GPT to succeed if glance
is new enough. If gpt is not in image_disk_formats then it will skip
on failure, else it will require that it works. This allows us to
merge this ahead of glance support, tolerate it passing once glance
lands that patch, and we can move this to always-require-if-configured
later.

Related to blueprint glance-as-defender

Change-Id: Idefcf5c67d40809377d72a87c0c3b87b0a36e097