Wait for VM ACTIVE state is added for fixed network case
When fixed network in tempest config is used for validation, IP
address will be returned in REST response to  get server api
only after VM will be in ACTIVE state. AS it mentioned in comment in
source  code:
"
NOTE(danms): We create these with no waiters because we will wait
for them to be validatable (i.e. SSHABLE) below. That way some of
the server creation overlap each other and with create_port.
"
But  wait_for_ssh_or_ping will call get_server_ip which fail for
server in building state (as there is no IP in server properties)

Finally we got KeyError: '<Network Name>'

Closes-Bug: 2086461
Change-Id: I21c5a81c41e771f7566bf19cac8e3b8f5f190b27