Clear pool for each new HTTP Connection
The memory consumption of Tempest can be high when HTTPS is used and
Tempest uses an older version of urllib3 as a dependency (1.26.x). This
is caused by Tempest, in conjunction with urllib3, loading certificates
into memory for each instance of RestClient and releasing them only at
the end of tests execution. The issue is particularly noticeable when
the certificate file is large.

This patch fixes this issue by clearing the pool in PoolManager after
each HTTP request.

Closes-Bug: #2088074
Change-Id: Iaede004522cd2e2386bac14ff725fa8e31d4991f