Add qcow2 format to upload image test
When we test uploading a volume to image, we have a config option
to specify the disk format. Currently the default disk format is
'raw' but most cinder backends store the volume in 'raw' format
and we end up avoiding the conversion workflow which becomes an
untested code path and could lead to bugs as recently discovered
with [1].
This patch adds the 'raw' and 'qcow2' as the default formats.
The config option is also updated to List so we have add multiple
formats to be tested by the same test.

[1] https://bugs.launchpad.net/cinder/+bug/2092534

Depends-On: https://review.opendev.org/c/openstack/cinder/+/938265

Related-Bug: #2092534

Change-Id: I5f0f005d9487d7cf59dfa818e2a327f6d3956d1b