Use stable constraint in tox to release new tag for 2025.1
We are going to release Tempest new tag to declare the start of
support for stable/2025.1. So that new tag use stable/2025.1
constraint in the tox env.

For example: if anyone use Tempest new tag in future say
1 year later then tox env also should use the 2025.1 contstraint for
compatibility instead master.

Also, making ironic-tempest-bios-ipmi-direct-tinyipa as
non voting to merge his change.

These need to be reverted once release is done.

Change-Id: I0df47c8c574c91b6c30fcc5dba40b83e2075b109