replace devstack-plugin-ceph-tempest-py3
this change replaces devstack-plugin-ceph-tempest-py3
with nova-ceph-multistore to add extra coverage of image
conversion and multi store.

devstack-plugin-ceph-tempest-py3 was disabled in gate 3 years
ago in Ibcaf1bce80521e05d76199ddebccb33488360644 as a temporary
measure. we never got around to re-enabling it so this change
also makes sure nova-ceph-multistore is run in gate to
have a ceph job in gate.

nova-ceph-multistore is removed from experimental as we do
not need the duplication since it is now in check

Related-Bug: #2092534
Change-Id: If0138b080a2222c95ae3c6d92221dfa1c8694db0