[Cinder] Preparation for supporting Cinder replication testing
This patch introduces a new configuration option `replication`
under the [volume-feature-enabled] group in Tempest. It enables
toggling of replication-related test execution.

It also adds support for the `failover_host` admin action in the
volume v3 services client, allowing tests to trigger failover
operations on Cinder backend hosts.

A waiter utility was added to wait for a volume to reach a specific
replication status. This will be used to validate replication state
transitions such as enabled and failed-over.

Unit tests have been added for both the new waiter function
and the service client method.

Release notes are included to document the new configuration option
and the service client/waiter additions.

This change is a preparatory step toward implementing volume
replication tests in the cinder-tempest-plugin repository.

This enables the Tempest framework to support upcoming Cinder
replication tests.

Change-Id: I0727e89a0f90d10d19e0e9ea16ebea3a3a90bc43
Signed-off-by: lkuchlan <lkuchlan@redhat.com>