Use next cidr to the configured one in the allowed_address_pair tests
In the
AllowedAddressPairTestJSON.test_update_port_with_cidr_address_pair API
test "self.cidr" was used as cidr to be set as allowed_address_pair for
the port. Since [1] it is forbidden by the ML2/OVN backend to set IP
address assigned to the metadata port (owner "network:distributed") as
allowed_address_pair for any port in the same network. This constraint
also applies to the CIDRs set as allowed_address_pair if metadata port's
IP address is included in that CIDR.
To avoid the issue in the tempest test, it now uses subnet which is next
to the one set as self.cidr - that way it will for sure not overlap with
the IP of the metadata port.

[1] https://review.opendev.org/c/openstack/neutron/+/954570

Related-bug: #2116249

Change-Id: Ib6f33a240ba78b39a28cb5aad6366cb29c26a568
Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>