New test case for Swift Policy quota limit added
This test verifies that quota limits are enforced per storage policy. It
sets a 10-byte quota on the "silver" policy and creates a container
using that policy. It then tries to upload an object larger than the
quota, expecting an OverLimit exception. Next, it creates a default
(non-policy) container and successfully uploads the same large object,
confirming that the quota enforcement is specific to the storage policy.
The test also ensures the response headers are correct after a
successful upload.

Closes-Bug: #2122673

Depends-On: https://review.opendev.org/c/openstack/devstack/+/955473

Change-Id: I3fb85118af7f62074556eaaf619dab3ddcc9be5b
Signed-off-by: Tanvi Nautiyal <tnautiya@redhat.com>