Use stable constraint in tox to release new tag for 2025.2
We are going to release Tempest new tag to declare the start of
2025.2 release support. So that new tag use stable/2025.2
constraint in the tox env.

For example: if anyone use Tempest new tag in future say
1 year later then tox env also should use the 2025.2 contstraint for
compatibility instead master.

Also, making ironic-tempest-bios-ipmi-direct-tinyipa as
non voting to merge his change.

These need to be reverted once release is done.

Change-Id: Id24617cfa114f1d5ab4694b69d9bb4e882adfb30
Signed-off-by: Ghanshyam Maan <gmaan@ghanshyammann.com>