Break out novnc connection validation to a mixin
Ironic is adding functional and integration testing for its
graphical console support and will need to do the same websocket
validation as this test. This change breaks out that validation into a
mixin class so that it can be used from ironic-tempest-plugin.

The first usage will be in the functional test with fake drivers [1] and
the second will be added to the existing baremetal basic ops[2]

[1] https://review.opendev.org/c/openstack/ironic-tempest-plugin/+/963152
[2] https://opendev.org/openstack/ironic-tempest-plugin/src/branch/master/ironic_tempest_plugin/tests/scenario/test_baremetal_basic_ops.py#L225

Change-Id: I27c3b574bac86cc8adda1944a1d843e0aec3d8e5
Signed-off-by: Steve Baker <sbaker@redhat.com>