Drop py3.9 from Tempest
Tempest python 3.9 unit test job is failing because requirement
repo dropped the python 3.9 constraints

- https://review.opendev.org/c/openstack/requirements/+/957587

We discussed this type of issue in past if requirement repo
should keep the pythion old version constraints

ref: https://lists.openstack.org/pipermail/openstack-discuss/2022-May/028711.html

Agreement to fix this is for tempest to drop the support of Python version.
Tempest should drop python3.9 support and if anyone wanted to test python3.9
env then that can be done by installing the tempest in python 3.10 (or
higher python version) env and test the python3.9 env.

Closes-Bug: #2130665

Change-Id: I3f21aa6c44779cc48aa175a1a801d59d125deeb5
Signed-off-by: Ghanshyam Maan <gmaan@ghanshyammann.com>