Use local variable for live migration server client
Avoid changing self.server_client to project manager permanently
because [compute_feature_enabled]live_migrate_back_and_forth might be
set to True. If it is, the test will live migrate the server back to
the source using the os-migrate-server:migrate_live:host API, which is
not allowed for project manager by default policy.

This allows use of the project manager or admin client on a per request
basis.

Change-Id: I0ca03bda1bc560d446827be2c95b8932106f7009
Signed-off-by: melanie witt <melwittt@gmail.com>